ForeachLoopExamples
This project is designed to learn how to use the foreach loop in C#. The foreach loop allows us to operate on all elements of a collection (such as an array or list) in sequence. This example shows how to use the foreach loop to traverse array and list elements.